add static optimization for form data
