+++ packages/web-component/src/lib/descope-wc/BaseDescopeWc.ts
@@ -35,17 +35,13 @@ import {
   FlowConfig,
 } from '../types';
 import initTemplate from './initTemplate';
+import {
+  extractNestedAttribute,
+  transformFlowInputFormData,
+} from '../helpers/flowInputs';
 
 // this is replaced in build time
 declare const BUILD_VERSION: string;
-
-const extractNestedAttribute = (
-  formData: Record<string, string | Record<string, string>>,
-  attr: string,
-) =>
-  Object.fromEntries(
-    Object.entries(formData).map(([name, values]) => [name, values[attr]]),
-  );
 
 // this base class is responsible for WC initialization
 class BaseDescopeWc extends HTMLElement {
@@ -77,6 +73,10 @@ class BaseDescopeWc extends HTMLElement {
   };
 
   #init = false;
+
+  formConfig = transformFlowInputFormData(this.getAttribute('form'));
+
+  formConfigValues = extractNestedAttribute(this.formConfig, 'value');
 
   loggerWrapper = {
     error: (message: string, description = '') => {
@@ -155,43 +155,6 @@ class BaseDescopeWc extends HTMLElement {
 
   get flowId() {
     return this.getAttribute('flow-id');
-  }
-
-  get form() {
-    try {
-      const form = (JSON.parse(this.getAttribute('form')) || {}) as Record<
-        string,
-        any
-      >;
-
-      let formData = form;
-      const vals = Object.values(form);
-
-      // transform values to object structure if needed
-      if (vals.some((s) => typeof s === 'string')) {
-        formData = Object.fromEntries(
-          Object.keys(form).map((key) =>
-            typeof form[key] !== 'string'
-              ? [key, form[key]]
-              : [key, { value: form[key] }],
-          ),
-        );
-      }
-
-      return Object.entries(formData).reduce(
-        (prev, [name, value]) => ({
-          ...prev,
-          [`form.${name}`]: value,
-        }),
-        formData,
-      );
-    } catch (e) {
-      return {};
-    }
-  }
-
-  get formConfigValues() {
-    return extractNestedAttribute(this.form, 'value');
   }
 
   get client() {
